pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yamldoc"]

[project]
name = "yaml-doc"
version = "1.0.0"
authors = [
    { name="Mizuki Hikaru", email="mizuki@hikaru.org" },
]
description = "yamldoc takes a YAML file and a Jinja2 template and renders a HTML or PDF file, populating the variables in the  Jinja2 template with data from your YAML file."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml",
    "jinja2",
    "weasyprint",
    "mistletoe",
]

[project.urls]
Homepage = "https://github.com/mizuki-hikaru/yamldoc"
Issues = "https://github.com/mizuki-hikaru/yamldoc/issues"

[project.scripts]
yamldoc = "yamldoc:cli"
